# doc/swagger.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: string
  /user:
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request.
      tags: &ref_0
        - User
    get:
      operationId: UsersController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
      tags: *ref_0
  /user/{id}:
    get:
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request.
        '404':
          description: Not found.
      tags: *ref_0
    put:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Not found.
      tags: *ref_0
    delete:
      operationId: UsersController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '404':
          description: Not found.
      tags: *ref_0
  /comment:
    post:
      operationId: CommentsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Bad request.
      tags: &ref_1
        - Comment
    get:
      operationId: CommentsController_findAll
      parameters:
        - name: articleId
          required: true
          in: query
          schema:
            type: string
        - name: page
          required: true
          in: query
          schema:
            type: number
        - name: limit
          required: true
          in: query
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
      tags: *ref_1
  /comment/{id}:
    get:
      operationId: CommentsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Bad request.
        '404':
          description: Not Found.
      tags: *ref_1
    delete:
      operationId: CommentsController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '404':
          description: Not Found.
      tags: *ref_1
  /article:
    post:
      operationId: ArticlesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          description: Bad request.
        '422':
          description: Unprocessed entity.
      tags: &ref_2
        - Article
    get:
      operationId: ArticlesController_findAll
      parameters:
        - name: status
          required: false
          in: query
          schema:
            enum:
              - draft
              - published
              - archived
            type: string
        - name: categoryId
          required: false
          in: query
          schema:
            type: string
        - name: tag
          required: false
          in: query
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
      tags: *ref_2
  /article/{id}:
    get:
      operationId: ArticlesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          description: Bad request.
        '404':
          description: Not Found.
      tags: *ref_2
    put:
      operationId: ArticlesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          description: Bad request.
        '404':
          description: Not found.
      tags: *ref_2
    delete:
      operationId: ArticlesController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '400':
          description: Bad request.
        '404':
          description: Not found.
      tags: *ref_2
  /category:
    post:
      operationId: CategoriesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad request.
      tags: &ref_3
        - Category
    get:
      operationId: CategoriesController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
      tags: *ref_3
  /category/{id}:
    get:
      operationId: CategoriesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad request.
        '404':
          description: Not found.
      tags: *ref_3
    put:
      operationId: CategoriesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad request.
        '404':
          description: Not Found.
      tags: *ref_3
    delete:
      operationId: CategoriesController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: ''
        '400':
          description: Bad request.
        '404':
          description: Not Found.
      tags: *ref_3
info:
  title: Knowledge Hub
  description: The Knowledge Hub API description
  version: '1.0'
  contact: {}
tags:
  - name: Knowledge Hub
    description: ''
servers: []
components:
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          description: User login
          example: user1
        password:
          type: string
          description: User password
          example: '123456'
        role:
          enum:
            - admin
            - editor
            - viewer
          type: string
          description: User role
          example: admin
      required:
        - login
        - password
    UserResponse:
      type: object
      properties:
        id:
          type: string
          description: User Id (UID)
          example: e30c4ac7-5be6-410a-ac61-c59e3eb5ea4f
        login:
          type: string
        role:
          enum:
            - admin
            - editor
            - viewer
          type: string
        createdAt:
          type: number
        updatedAt:
          type: number
      required:
        - id
        - login
        - role
        - createdAt
        - updatedAt
    UpdateUserDto:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
      required:
        - oldPassword
        - newPassword
    CreateCommentDto:
      type: object
      properties:
        content:
          type: string
          description: Comment content
          example: Great Article
        articleId:
          type: string
          description: articleId
          example: 61ff1183-ef8f-4e21-b9e5-01f7463ad139
        authorId:
          type: string
          nullable: true
          description: authorId
          example: 61ff1183-ef8f-4e21-b9e5-01f7463ad139
      required:
        - content
        - articleId
    Comment:
      type: object
      properties:
        id:
          type: string
          description: User Id (UID)
          example: e30c4ac7-5be6-410a-ac61-c59e3eb5ea4f
        content:
          type: string
        articleId:
          type: string
        authorId:
          type: string
          nullable: true
        createdAt:
          type: number
      required:
        - id
        - content
        - articleId
        - authorId
        - createdAt
    CreateArticleDto:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        status:
          enum: &ref_4
            - draft
            - published
            - archived
          type: string
        authorId:
          type: string
          nullable: true
        categoryId:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - content
    Article:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        status:
          enum:
            - draft
            - published
            - archived
          type: string
        authorId:
          type: string
          nullable: true
        categoryId:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: number
        updatedAt:
          type: number
      required:
        - id
        - title
        - content
        - authorId
        - categoryId
        - createdAt
        - updatedAt
    UpdateArticleDto:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        status:
          enum: *ref_4
          type: string
        authorId:
          type: string
          nullable: true
        categoryId:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - content
    CreateCategoryDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
        - description
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
      required:
        - id
        - name
        - description
    UpdateCategoryDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
        - description